scripts/process_sales/pyproject.toml
[project]
name = "ubiquus-sales-processor"
version = "0.1.0"
description = "Monthly sales processor for Ubiquus invoicing"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "google-auth>=2.23.0",
    "google-api-python-client>=2.100.0",
]

[project.scripts]
process-monthly-sales = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
